add overflow

+++ lib/src/rviews/edge.dart
@@ -1,10 +1,10 @@
 // This is used for implemented repert
 part of 'rview_bases.dart';
 abstract class EdgInset {
-  final int top;
-  final int left;
-  final int right;
-  final int bottom;
+  final double top;
+  final double left;
+  final double right;
+  final double bottom;
 
   const EdgInset(
       {this.top = 0, this.bottom = 0, this.left = 0, this.right = 0});
@@ -13,6 +13,6 @@ abstract class EdgInset {
 class REdgetInset extends EdgInset {
   static const zero = REdgetInset();
   const REdgetInset({super.top, super.bottom, super.left, super.right});
-  const REdgetInset.all(int value)
+  const REdgetInset.all(double value)
       : super(top: value, bottom: value, left: value, right: value);
 }
